+++ src/frontend/src/controllers/API/queries/folders/index.tsx
@@ -1,1 +1,2 @@
+export * from "./use-delete-folders";
 export * from "./use-post-upload-folders";
