+++ src/frontend/src/pages/MainPage/pages/mainPage/index.tsx
@@ -1,3 +1,5 @@
+import { useDeleteFolders } from "@/controllers/API/queries/folders";
+import useAlertStore from "@/stores/alertStore";
 import { useEffect, useState } from "react";
 import { Outlet, useLocation, useNavigate } from "react-router-dom";
 import DropdownButton from "../../../../components/dropdownButtonComponent";
@@ -10,8 +12,8 @@ import {
 import useFlowsManagerStore from "../../../../stores/flowsManagerStore";
 import { useFolderStore } from "../../../../stores/foldersStore";
 import ModalsComponent from "../../components/modalsComponent";
-import useDeleteFolder from "../../hooks/use-delete-folder";
 import useDropdownOptions from "../../hooks/use-dropdown-options";
+import { getFolderById } from "../../services";
 
 export default function HomePage(): JSX.Element {
   const uploadFlow = useFlowsManagerStore((state) => state.uploadFlow);
@@ -28,6 +30,12 @@ export default function HomePage(): JSX.Element {
   const setFolderToEdit = useFolderStore((state) => state.setFolderToEdit);
   const navigate = useNavigate();
 
+  const setSuccessData = useAlertStore((state) => state.setSuccessData);
+  const setErrorData = useAlertStore((state) => state.setErrorData);
+  const folderToEdit = useFolderStore((state) => state.folderToEdit);
+  const myCollectionId = useFolderStore((state) => state.myCollectionId);
+  const getFoldersApi = useFolderStore((state) => state.getFoldersApi);
+
   useEffect(() => {
     setCurrentFlowId("");
   }, [pathname]);
@@ -38,7 +46,31 @@ export default function HomePage(): JSX.Element {
     is_component,
   });
 
-  const { handleDeleteFolder } = useDeleteFolder({ navigate });
+  const { mutate } = useDeleteFolders();
+
+  const handleDeleteFolder = () => {
+    mutate(
+      {
+        folder_id: folderToEdit?.id!,
+      },
+      {
+        onSuccess: () => {
+          setSuccessData({
+            title: "Folder deleted successfully.",
+          });
+          getFolderById(myCollectionId!);
+          getFoldersApi(true);
+          navigate("/all");
+        },
+        onError: (err) => {
+          console.error(err);
+          setErrorData({
+            title: "Error deleting folder.",
+          });
+        },
+      },
+    );
+  };
 
   return (
     <>
